requirements.txt
asgiref==3.2.5
Django==3.0.4
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.1
phonenumbers
django-phonenumber-field
